+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 0.1.6 (Unreleased)
 
 FEATURES:
- - `prodvana_release_channel` - added ReleaseChannelStable and ManualApproval precondition support
+ - `prodvana_release_channel`
+   - added ReleaseChannelStable and ManualApproval precondition support
+   - Added Protections support
 
 ## 0.1.5